Atualizando scripts do banco de dados.

+++ assets/database/database_creation_script.sql
@@ -202,8 +202,8 @@ CREATE TABLE `visit` (
   `user_id` bigint(20) NOT NULL,
   `datetime` datetime NOT NULL,
   `carried_out` tinyint(1) NOT NULL DEFAULT 0,
-  `visit_rating` tinyint(1) unsigned DEFAULT NULL,
-  `property_rating` tinyint(1) unsigned DEFAULT NULL,
+  `visit_rating` int(1) DEFAULT NULL,
+  `property_rating` int(1) DEFAULT NULL,
   `comments` text DEFAULT NULL,
   `created_at` timestamp NOT NULL DEFAULT current_timestamp(),
   PRIMARY KEY (`id`),
@@ -217,14 +217,3 @@ CREATE TABLE `visit` (
 --
 -- Dumping routines for database 'morada'
 --
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2023-11-26 21:44:43
